Changed validation to be a seperate worker not on the main thread (#3988)
* Changed validation to be a seperate worker not on the main thread

* Removed console log
