+++ server/routes/establishments/bulkUpload/report.js
@@ -200,7 +200,7 @@ const reportGet = async (req, res) => {
 const { acquireLock } = require('./lock');
 const router = require('express').Router();
 
-router.route('/:reportType').get(acquireLock.bind(null, reportGet, buStates.DOWNLOADING));
+router.route('/:reportType').get(acquireLock.bind(null, reportGet, buStates.DOWNLOADING, true));
 
 module.exports = router;
 module.exports.printLine = printLine;
